% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_polygons_df.R
\name{create_random_polygons_df}
\alias{create_random_polygons_df}
\title{Create random overlapping polygons}
\usage{
create_random_polygons_df(
  ext,
  n = 5,
  nmap_min = 50,
  nmap_max = 200,
  year_min = 1800,
  year_max = 1899,
  n.pnts = 10,
  progress = TRUE
)
}
\arguments{
\item{ext}{An sf object, within which the polygons will be sampled}

\item{n}{Integer, the number of groups to create}

\item{nmap_min}{Integer, minimum number of features within a group}

\item{nmap_max}{Integer, maximum number of features within a group}

\item{year_min}{Integer, minimum year}

\item{year_max}{Integer, maximum year}

\item{n.pnts}{Integer, number of points to sample for each feature to create a polygon}

\item{progress}{Logical, whether to show progress bar}
}
\value{
An sf dataframe of random overlapping polygons
}
\description{
Create random overlapping polygons
}
