% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2grid.R
\name{data2grid}
\alias{data2grid}
\title{Match covariates to raster data.}
\usage{
data2grid(x, covars)
}
\arguments{
\item{x}{Either an object of class \code{SpatRaster} or a list object returned from \link{get_grid}}

\item{covars}{Character vector specifying the covariates to add. Default is the full list,
i.e. \code{covars = c("area", "terrain", "malaria", "coast")} (see details).}
}
\value{
Returns either a dataframe with rows corresponding to grid cells (if \code{x} is a \code{SpatRaster} object) or
a the list object with the covariates matched to the data item (if \code{x} is the list object returned from
\link{get_grid}). The following covariates are matched by default:
\item{\code{area}}{Area of the grid cells in km2}
\item{\code{terrain}}{Average elevation in meters and average terrain ruggedness within each grid cell.
Requires installation of the package \code{elevatr}. Source: }
\item{\code{malaria}}{Mean temperature suitability index for malaria within each grid cell.
Requires installation of the package \code{malariaAtlas}. Source: }
\item{\code{coast}}{Distance to nearest coast in meters. Requires installation of the package
\code{rnaturalearthdata}. Source: }
}
\description{
Adds a specified list of common covariates to the gridded data frame. The function
takes either a SpatRaster object, in which case it returns a data frame with the
specified covariates, or a list object returned from \link{get_grid}, in which case it
returns the same list object with covariates matched to the data frame.
}
