% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2poly.R
\name{data2poly}
\alias{data2poly}
\title{Match covariates to polygon data.}
\usage{
data2poly(shp, covars)
}
\arguments{
\item{shp}{\code{sf} polygons data.frame, typically the output from \link{get_contours}.}

\item{covars}{Character vector specifying the covariates to add. Default is the full list,
i.e. \code{covars = c("area", "terrain", "malaria", "coast")} (see details).}
}
\value{
Returns a dataframe similar to \code{shp} with columsn added for covariates.
The following covariates are matched by default:
\item{\code{area}}{Area of the polygon in km2}
\item{\code{terrain}}{Average elevation in meters and average terrain ruggedness within each polygon.
Requires installation of the package \code{elevatr}. Source: }
\item{\code{malaria}}{Mean temperature suitability index for malaria within each polygon.
Requires installation of the package \code{malariaAtlas}. Source: }
\item{\code{coast}}{Distance to nearest coast from the border of polygons (in km).
Requires installation of the package \code{rnaturalearthdata}. Source: }
}
\description{
Adds a specified list of common covariates to an \code{sf} polygons data.frame. The function
takes an \code{sf} dataframe (typically the output from \link{get_contours}) and returns the same
data frame with the specified selection of covariates.
}
