% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_errors.R
\name{detect_errors}
\alias{detect_errors}
\title{Detect potential errors in maps}
\usage{
detect_errors(
  shp,
  id_var,
  period_var,
  errors,
  dist,
  parallel = TRUE,
  ncores,
  report = FALSE
)
}
\arguments{
\item{shp}{An sf dataframe containing all geometries}

\item{id_var}{Name of grouping variable (e.g., COWID)}

\item{period_var}{Name of year variable}

\item{errors}{Character vector of which errors to search for. See details.}

\item{dist}{Minimum allowable distance (in km) for which capitals can be outside polygons.}

\item{parallel}{Whether to use parallel processing. Default is TRUE.}

\item{ncores}{Number of cores to use for parallel processing. Default is to use all available cores.}

\item{report}{Whether the output should be a report of potential errors. Default is an sf dataframe
with columns for each error detection.}
}
\value{
Data frame
\item{\code{singles}}{Whether there are COWIDs with only one map (may indicate wrong COWID).}
\item{\code{year}}{Whether there are polygons with years outside the study window (i.e. 1750-1920) or if year is missing.}
\item{\code{overlap}}{Whether there are COWIDs for which one or more polygons do not overlap with any of the other polygons.}
\item{\code{capital}}{Whether any of the capitals associated with a COWID falls outside ALL polygons. The option \code{dist} can be
used to specify a minimum allowable distance.}
}
\description{
Detect potential errors in maps
}
