% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df}
\alias{example_df}
\title{Example data set}
\format{
\subsection{\code{example_df}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{id}{ID number of fictional country}
\item{name}{Name of fictional country}
\item{year}{Year of territorial extent}
...
}
}
}
\usage{
example_df
}
\description{
A data set of grouped polygons (by id and year) resembling the data structure
the package is designed to handle. The polygons are generated randomly and do not
reflect any actual political entities (historical or contemporary).
}
\keyword{datasets}
