% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\title{Merge LEGACIES data with covariates}
\usage{
get_covariates(x, vars, path, progress = TRUE, parallel = FALSE, coords)
}
\arguments{
\item{x}{List with named elements \code{r}, object of class \code{SpatRaster}, and \code{data},
object of class \code{data.frame}, OR a named object of class \code{data.frame}. The list object
is returned by the \code{get_grid} function if \code{return = "list"}.}

\item{vars}{Covariate data to merge. Default includes all variables. See details below.}

\item{path}{Character string specifying the path name for the LEGACIES team folder. Depends
on the user's computer, but generally should be along the lines of
\code{"~/user/OneDrive-NTNU/legacies-project"}.}

\item{progress}{Logical, whether progress updates should be displayed.}

\item{parallel}{Logical, whether to parallelize jobs (CURRENTLY NOT USED)}

\item{coords}{Character vector of length two specifying the names of the coordinate
(longitude/latitude) variables. Default is \code{c("lon", "lat")} which follows the
convention in \code{get_grid}.}
}
\value{
Data frame or list of data frame and SpatRaster. By default, all covariates are included,
but it is also possible to specify which covariates to include with the argument \code{vars},
which takes a character vector as input. The following data are available:
\item{\code{terrain}}{.}
\item{\code{climate}}{.}
\item{\code{crops}}{.}
\item{\code{rivers}}{.}
\item{\code{coast}}{}
}
\description{
Merge LEGACIES data with covariates
}
