% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\title{Create grid summarizing overlapping polygons.}
\usage{
get_grid(
  shp,
  ras,
  raster_from_shp = F,
  res = 1/2,
  by_grp = T,
  id_var,
  updates = T,
  return = c("list", "data")
)
}
\arguments{
\item{shp}{Shape}

\item{ras}{Raster}

\item{raster_from_shp}{Logical}

\item{res}{Resolution}

\item{by_grp}{By group}

\item{id_var}{ID variable name}

\item{updates}{Logical}

\item{return}{List or data.frame}
}
\value{
Data frame of grid cell-by-state (or grid cell-by-state-by-year).
}
\description{
Create gridded data by group (or group and year) summarizing the number of
polygons covering a given grid cell.
}
