% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_id_contour.R
\name{match_id_contour}
\alias{match_id_contour}
\title{Match contour polygons to ID data}
\usage{
match_id_contour(
  id_data,
  contour_data,
  contour_level = 0.25,
  multiple_levels = FALSE,
  covar_path
)
}
\arguments{
\item{id_data}{A dataframe with ID data (after being processed with the \code{parepare_id} function.)}

\item{contour_data}{An sf dataframe with contour polygons (from \code{get_contours}).}

\item{contour_level}{The probability level to use if matching to a single polygon per polity-year. Default is \code{0.25}.}

\item{multiple_levels}{Whether to include all contour levels, in which case there will be multiple rows per polity-year.}

\item{covar_path}{File path to folder containing HYDE population data. Typically the path will be something like \code{"~/user/OneDrive-NTNU/legacies-project"}.}
}
\value{
A dataframe with polity-year observations and columns for land area (based on contour polygons) and decade population estimates
based on HYDE population data (aggregated to contour polygons).
}
\description{
The function attaches contour polygon geometries to the ID data as well as data on land area and population.
}
