% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_id.R
\name{prepare_id}
\alias{prepare_id}
\title{Prepare ID data}
\usage{
prepare_id(data, multiple_capital = FALSE)
}
\arguments{
\item{data}{Raw ID data to be cleaned and expanded. Typically called \code{legacies_id_coding.xlsx}.}

\item{multiple_capital}{Logical, whether to keep all capitals if there are multiple capitals in
the same polity-year. In that case, a polity-year will be expanded to one row per capital.
Default is \code{FALSE}, in which case only the capital with the longest spell will be included.
If two or more capitals have the same spell, one capital will be selected at random.}
}
\value{
A data frame with polity-year observations and the variables decribed below:
\item{\code{polity_name}}{Name of polity}
\item{\code{COWID}}{COW code for the polity}
\item{\code{COWNUM}}{Numeric COW code for the polity}
\item{\code{year}}{Year of observation}
\item{\code{other_names}}{Alternative polity names}
\item{\code{start_year}}{Start year of a given polity spell}
\item{\code{end_year}}{End year of a given polity spell}
\item{\code{population_10k}}{Whether the polity-year lives up to the population criterium}
\item{\code{autonomy}}{Whether the polity-year is regarded as autonomous}
\item{\code{capital_name}}{Name of capital of the polity in a given year. If there are multiple capitals,
the most prominent capital is recorded here.}
\item{\code{capital_lon}}{Longitude of capital city}
\item{\code{capital_lat}}{Latitude of capital city}
\item{\code{capital_raw}}{Raw capital city information. Can contain information for multiple capital cities.
The capital variables above contain information on only the most prominent capital.}
\item{\code{hie_COWID}}{COWID for which the given polity is in a hierarchical relationship.
Can contain multiple COWIDs if the polity is in multiple hierarchical relationships.}
\item{\code{hie_type}}{Type of hierarchy (\code{1 = tributary}, \code{2 = dependency}). Can contain multiple.}
\item{\code{hie_tributary}}{Dummy equal to \code{1} if the polity is in at least one tributary relationship.}
\item{\code{hie_dependency}}{Dummy equal to \code{1} if the polity is in at least one dependency relationship.}
\item{\code{independent_dependency}}{Dummy equal to \code{1} if the polity is independent, i.e. it is not a dependency.}
\item{\code{independent_tributary}}{Dummy equal to \code{1} if the polity is independent, i.e. it is not a tributary.}
\item{\code{independent}}{Dummy equal to \code{1} if the polity is independent, i.e. neither dependency nor tributary.}
\item{\code{hie_raw}}{Raw hierarchy coding.}
\item{\code{ethnonational_group}}{}
\item{\code{EPR_link}}{}
\item{\code{ACD_actor_link}}{}
\item{\code{end_agent}}{}
\item{\code{destination_states}}{}
\item{\code{destination_state_1-5}}{}
}
\description{
The function prepares the raw ID data for further use by, e.g., expanding rows into
polity-year observations and extracting information on capitals and hierarchies.
}
