% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shapes.R
\name{prepare_shapes}
\alias{prepare_shapes}
\title{Prepare shapefiles}
\usage{
prepare_shapes(
  shp,
  state_data,
  id_var,
  period_var,
  range_min,
  range_max,
  fix_year = TRUE,
  year_na = TRUE,
  expand_range = TRUE,
  match_capitals = TRUE,
  exclude_hierarchy = c("none", "tributary", "dependency", "all"),
  exclude_core = TRUE,
  exclude_incomplete = TRUE,
  exclude_sovereign = TRUE,
  margin_sovereign = 5,
  crop_to_land = TRUE
)
}
\arguments{
\item{shp}{sf dataframe}

\item{state_data}{dataframe with state ID data}

\item{id_var}{ID variable, must be the same in both \code{shp} and \code{state_data}}

\item{period_var}{Year variable, must be the same in both \code{shp} and \code{state_data}}

\item{range_min}{Name of variable for lower year interval}

\item{range_max}{Name of variable for upper year interval}

\item{fix_year}{Logical, whether to fix three-digit years. Default is \code{TRUE}. See details.}

\item{year_na}{Logical, whether to exclude maps with no year indication (exact or range)}

\item{expand_range}{Logical, whether to expand rows for maps with year interval, default is \code{TRUE.} See details.}

\item{match_capitals}{Logical, whether to add information on capital cities to the \code{shp} data. Default is \code{TRUE}. See details.}

\item{exclude_hierarchy}{Character vector, whether to exclude maps for states that are subordinated (tributary or dependency). \code{none} (default)
will keep all maps regardless of hierarchy status, \code{tributary} will exclude tributary states (\code{hierarchy == 1}), \code{dependency} will exclude
dependencies (\code{hierarchy == 2}), and \code{both} will exclude both tributaries and dependencies. See also \code{combine_hierarchy}.}

\item{exclude_core}{Exclude maps coded as core (i.e., include only the complete shape of both core and periphery).
Default is \code{TRUE}.}

\item{exclude_incomplete}{Exclude maps with incomplete borders. Default is \code{TRUE}.}

\item{exclude_sovereign}{Exclude maps for states that are not sovereign at the time of the map. See details.}

\item{margin_sovereign}{Number of years before or after a state is sovereign that a map is still included (if \code{exclude_sovereign} is \code{TRUE}). Default is 5 years.}

\item{crop_to_land}{Logical, whether to crop geometries to land (default is \code{TRUE})}
}
\value{
An sf dataframe.
Further details:
\item{\code{expand_year}}{Expands rows for maps with interval range rather than specific year. The lower and
upper limits are always retained. In between the limits, rows are only in ten-year interval.}
\item{\code{match_capitals}}{Capital coordinates are stored as separate geometry. If there are multiple capitals
in the same year, all of them withh be included. The capital names will then be separated by semicolon.}
\item{\code{exclude_sovereign}}{Excludes maps for states that are not sovereign. If the year of the map is within
5 years of a state being sovereign, the map is still included.}
}
\description{
The function prepares data for further processing with other functions in the `legaciesr' package,
e.g., expands rows for maps coded with a range (rather than specific year), excludes certain maps, etc.
}
