% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_polygons.R
\name{contour_polygons}
\alias{contour_polygons}
\title{Create Contour Polygons}
\usage{
contour_polygons(
  shp,
  cuts = 10,
  smooth_method = c("ksmooth", "chaikin", "spline"),
  smoothness_factor = 2,
  res = 1/60,
  invalid_geom = c("none", "exclude", "fix_exclude", "fix_all"),
  keep_vars = F
)
}
\arguments{
\item{shp}{The SF PolygonsDataFrame to be aggregated. Must consist of more than one feature.}

\item{cuts}{An integer of length 1 specifying the number of (equally spaced) contour polygons to be returned.
For example, if \code{cuts = 10} the function will return 10 polygons representing the 10 deciles (0-0.1,0.1-0.2,...).}

\item{smooth_method}{Character scalar specifying the smoothing method used to convert raster to polygon.
The options are \code{"ksmooth"} (default), \code{"chaikin"}, and \code{"spline"}.}

\item{smoothness_factor}{Numerical scalar indicating the amount of smoothing the \code{\link[smoothr]{smooth}} function should do.
Default is 2.}

\item{res}{Resolution of the raster used to calculate polygon density. Higher resolution (i.e., lower numbers)
creates smoother borders but also increases processing time. Default is 1/60 degrees.}

\item{invalid_geom}{One of "none" (default), "exclude", "fix_exclude", or "fix_all". The function attempts to rebuild invalid geometries
using sf::st_make_valid(shp). However, some geometries may be invalid even after attempting to rebuild. If "none" the function will
return an error in case of invalid geometries (after attempting to rebuild). If "exclude", any invalid geometries will be removed.
If "fix_exclude", invalid geometries will be rebuilt using default arguments and any remaining invalid
geometries will be removed. If "fix_all", the function will attempt to rebuild invalid geometries by
iteratively lowering the snappingnprecision (in \code{s2::s2_snap_precision()}). It will return an error
if it fails to rebuild valid geometries. The use of "fix_all" is experimental and the result should be
inspected manually.}

\item{keep_vars}{logical, indicating is the returned dataframe should include columns from the input dataframe.
Since the number of rows do not match, only the first row from the input dataframe is used across all
rows in the output.}
}
\value{
Returns an SF PolygonsDataFrame with the same number of features as specified in *cuts*. The density,
 or percentile, is stored in the attributes \code{"prob"} and \code{"label"}.
}
\description{
This function creates contour polygons from multiple overlapping polygons.
}
