% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_errors.R
\name{detect_errors}
\alias{detect_errors}
\title{Detect potential errors in maps}
\usage{
detect_errors(
  shp,
  capital_data,
  id_var,
  period_var,
  errors,
  dist,
  parallel = TRUE,
  ncores,
  progress = TRUE,
  report = TRUE,
  returnList = TRUE
)
}
\arguments{
\item{shp}{An sf dataframe containing all geometries}

\item{capital_data}{Dataframe containing information on capitals (by year). Should have the same
\code{id_var} and \code{period_var} names as \code{shp}.}

\item{id_var}{Name of grouping variable (e.g., COWID)}

\item{period_var}{Name of year variable}

\item{errors}{Character vector of which errors to search for. See details.}

\item{dist}{Minimum allowable distance (in km) for which capitals can be outside polygons.}

\item{parallel}{Whether to use parallel processing. Default is TRUE.}

\item{ncores}{Number of cores to use for parallel processing. Default is to use all available cores.}

\item{progress}{Whether to print progress updates. Default is TRUE.}

\item{report}{Whether the output should be a report of potential errors. Default is an sf dataframe
with columns for each error detection.}

\item{returnList}{Whether to return a list containing the full sf data frame AND separate data frames
for each error type (default) or just the full sf data frame. The full sf data frame will either way
include columns indicating potential errors for each type.}
}
\value{
Data frame
\item{\code{id_duplicates}}{Whether COWID and COWNUM do not align, i.e. are there duplicate COWIDs with different COWNUMs?}
\item{\code{missing_id}}{Whether there are geometries with missing \code{id_var}.}
\item{\code{singles}}{Whether there are COWIDs with only one map (may indicate wrong COWID).}
\item{\code{year}}{Whether there are polygons with years outside the study window (i.e. 1750-1920) or if year is missing.}
\item{\code{overlap}}{Whether there are COWIDs for which one or more polygons do not overlap with any of the other polygons.}
\item{\code{capital}}{Whether any of the capitals associated with a COWID falls outside ALL polygons. The option \code{dist} can be
used to specify a minimum allowable distance.}
\item{\code{empty}}{Check whether there are empty geometries.}
}
\description{
Detect potential errors in maps
}
