% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contours.R
\name{get_contours}
\alias{get_contours}
\title{Get Contour Polygons Across Groups}
\usage{
get_contours(
  shp,
  grp_id,
  by_period = FALSE,
  interval,
  period_name,
  nmap_threshold = 5,
  returnList = F,
  progress = F,
  parallel = F,
  ncores,
  ...
)
}
\arguments{
\item{shp}{an sf dataframe containing all geometries (polygons).}

\item{grp_id}{grouping variable.}

\item{by_period}{logical, indicating if shapes should be divided into groups by time in addition to IDs.
Default is \code{FALSE}.}

\item{interval}{scalar or vector indicating the intervals to group by. If \code{NULL} (default), the
intervals will automatically be set 20 and include the full range (i.e. \code{seq(min, max, 20)}).
For user specification, the value can be either a scalar, indicating the length of intervals
(in which case the full range is used), or a vector specifying the exact breaks, including both the
minimum and maximum.}

\item{period_name}{name of the period variable to be used for time grouping.
This argument is necessary if \code{by_period} is \code{TRUE}.}

\item{nmap_threshold}{scalar indicating the number of shapes requires within each group. Default is 5.}

\item{returnList}{logical, whether to return a list of geometries by group.
Default is to return an sf dataframe containing all geometries.}

\item{progress}{logical, indicating whether to show a progress bar.}

\item{parallel}{logical, indicating whether to use parallel processing.}

\item{ncores}{scalar, indicating the number of cores to use for parallel processing.
Default is the number of cores available minus 1.}

\item{...}{Additional arguments passed to \code{\link{contour_polygons}}.}
}
\value{
Returns either an sf dataframe (default) or a list of sf dataframes (one list item per group).
}
\description{
This function creates contour polygons across multiple groups, either by group alone or
in combination with a time column.
}
