% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shapes.R
\name{prepare_shapes}
\alias{prepare_shapes}
\title{Prepare shapefiles}
\usage{
prepare_shapes(
  shp,
  state_data,
  id_var,
  period_var,
  range_min,
  range_max,
  fix_year = TRUE,
  year_na = TRUE,
  expand_range = 20L,
  match_capitals = TRUE,
  exclude_source_year = NA,
  exclude_core = TRUE,
  exclude_incomplete = TRUE,
  exclude_sovereign = TRUE,
  margin_sovereign = 5,
  crop_to_land = TRUE
)
}
\arguments{
\item{shp}{sf dataframe}

\item{state_data}{dataframe with state ID data}

\item{id_var}{ID variable, must be the same in both \code{shp} and \code{state_data}}

\item{period_var}{Year variable, must be the same in both \code{shp} and \code{state_data}}

\item{range_min}{Name of variable for lower year interval}

\item{range_max}{Name of variable for upper year interval}

\item{fix_year}{Logical, whether to fix three-digit years. Default is \code{TRUE}. See details.}

\item{year_na}{Logical, whether to exclude maps with no year indication (exact or range)}

\item{expand_range}{Integer, indicating for maps with year interval (rather than specific year) the number of
years between repeating maps. If \code{0}, only the start and end date will be included. Otherwise, the specified
number indicates the how often maps should be repeated. Default is \code{20} years.}

\item{match_capitals}{Logical, whether to add information on capital cities to the \code{shp} data. Default is \code{TRUE}. See details.}

\item{exclude_source_year}{Integer, indicating if maps from sources before a given year should be dropped. Default is
\code{NA}, which means all maps, regardless of source year, are included.}

\item{exclude_core}{Exclude maps coded as core (i.e., include only the complete shape of both core and periphery).
Default is \code{TRUE}.}

\item{exclude_incomplete}{Exclude maps with incomplete borders. Default is \code{TRUE}.}

\item{exclude_sovereign}{Exclude maps for states that are not sovereign at the time of the map.
Sovereignty is defined by the state-years in the \code{state_data} provided. If, for instance, one wishes to include only "states" and
"quasi-states" then the \code{state_data} needs to be subset to those categories. See also details.}

\item{margin_sovereign}{Number of years before or after a state is sovereign that a map is still included (if \code{exclude_sovereign} is \code{TRUE}). Default is 5 years.}

\item{crop_to_land}{Logical, whether to crop geometries to land (default is \code{TRUE})}
}
\value{
An sf dataframe.
Further details:
\item{\code{expand_range}}{Expands rows for maps with interval range rather than specific year. The lower and
upper limits are always retained. In between the limits, rows are only in the specific interval. For example
if \code{expand_range = 20}, years divisible by 20 will be included (in addition to the lower and upper limits).}
\item{\code{match_capitals}}{Capital coordinates are stored as separate geometry. If there are multiple capitals
in the same year, all of them withh be included. The capital names will then be separated by semicolon.}
\item{\code{exclude_sovereign}}{Excludes maps for states that are not sovereign. If the year of the map is within
5 years of a state being sovereign, the map is still included.}
}
\description{
The function prepares data for further processing with other functions in the `legaciesr' package,
e.g., expands rows for maps coded with a range (rather than specific year), excludes certain maps, etc.
}
